Edit

+++ src/main/java/com/capstone/joonggo/dto/MarketDto.java
@@ -10,13 +10,13 @@ public class MarketDto {
 
     private String title;
 
-    private Long price;
+    private int price;
 
     private Long postId;
 
     private LocalDateTime createdTime;
 
-    public MarketDto(String title, Long price,Long postId, LocalDateTime createdTime) {
+    public MarketDto(String title, int price,Long postId, LocalDateTime createdTime) {
         this.title = title;
         this.price = price;
         this.postId = postId;
